Refactor verification functions
