+++ src/utils/GetCookie.ts
@@ -1,0 +1,1 @@
+export type GetCookie = (name: string) => string | undefined
