+++ src/utils/SetCookie.ts
@@ -1,0 +1,1 @@
+export type SetCookie = (name: string, value: string) => void
