+++ src/utils/cookie/index.ts
@@ -1,2 +1,4 @@
 export * from './getCookie'
 export * from './setCookie'
+export * from './getSignedCookie'
+export * from './setSignedCookie'
