+++ src/utils/signData/signData.test.ts
@@ -1,5 +1,5 @@
-import { signData } from './sign-data'
-import { verifyData } from './verify-data'
+import { signData } from './signData'
+import { verifyData } from './verifyData'
 
 const testSecret =
   'fkxAHP5whEOjjJh4SFvYvQ9BiqBc8DMqQiX4MMFOcSUx5Qh5xxOI2wqQMRfK53aTOyc5RyEimYQBsA7lWu9kag=='
