+++ src/utils/verifyData/verifyData.test.ts
@@ -1,6 +1,6 @@
-import { verifyData } from './verify-data'
+import { verifyData } from './verifyData'
 import jwt from 'jsonwebtoken'
-import { signData } from './sign-data'
+import { signData } from './signData'
 
 const testSecret =
   'fkxAHP5whEOjjJh4SFvYvQ9BiqBc8DMqQiX4MMFOcSUx5Qh5xxOI2wqQMRfK53aTOyc5RyEimYQBsA7lWu9kag=='
